feat(ui): add Magic UI Blur Fade via shadcn registry
Install motion and official @magicui/blur-fade into src/components/ui;
fix shadcn aliases to src paths; wire homepage and downloads/pricing imports.
Include app reduced-motion and outputsShowcaseFigure reveal targets.

Made-with: Cursor
