+++ package.json
@@ -34,6 +34,7 @@
     "clsx": "^1.2.1",
     "docusaurus-plugin-sass": "^0.2.4",
     "framer-motion": "^11.14.4",
+    "motion": "^12.38.0",
     "prism-react-renderer": "^2.1.0",
     "react": "^18.2.0",
     "react-cookie-manager": "^4.0.3",
