+++ src/components/HomepageDownloads/index.tsx
@@ -4,7 +4,7 @@ import { translate } from "@docusaurus/Translate";
 import { DownloadUrl } from "@site/src/download_url";
 import React from "react";
 
-import { BlurFade } from "../magic-ui/BlurFade";
+import { BlurFade } from "@site/src/components/ui/blur-fade";
 import { Button } from "../ui/button";
 
 type DownloadBtnProps = {
